moves decryption/validation to anonymization.go
