+++ clients/go/apiban/ip.go
@@ -4,51 +4,8 @@ import (
 	"fmt"
 	"log"
 	"strconv"
-	"strings"
 	"time"
-
-	"github.com/intuitivelabs/anonymization"
 )
-
-func decryptIp(encrypted string, kvCode interface{}) (decrypted string, err error) {
-	// check the string type for "encrypt" field
-	err = nil
-	decrypted = encrypted
-	kvCodeStr, ok := kvCode.(string)
-	if !ok {
-		err = ErrJsonEncryptFieldNotString
-		return
-	}
-	log.Print("encrypt field: ", kvCodeStr)
-	if isPlainTxt(kvCodeStr) {
-		// not encrypted; passthrough
-		return
-	}
-	// check if encrypt flags are part of the encrypt field
-	split := strings.Split(kvCodeStr, "|")
-	switch len(split) {
-	case 1:
-		// flags are not there; nothing special to do
-	case 2:
-		// flags are present; get the code
-		kvCodeStr = split[1]
-		log.Print("key validation code: ", kvCodeStr)
-	default:
-		// broken format
-		err = fmt.Errorf("encrypt field unknown format: %s", kvCodeStr)
-		return
-	}
-	if (Validator == nil) || (Ipcipher == nil) {
-		err = ErrEncryptNoKey
-		return
-	}
-	if !Validator.Validate(kvCodeStr) {
-		err = ErrEncryptWrongKey
-		return
-	}
-	decrypted = Ipcipher.(*anonymization.Ipcipher).DecryptStr(encrypted)
-	return
-}
 
 // IP Resource JSON objects in API responses.
 // It represents IPs of blocked/allowed IP addresses
@@ -90,10 +47,19 @@ func (ips IpVector) Blacklist(ttl time.Duration) error {
 }
 
 func (ip *IP) Process(ttl time.Duration, api APICode) error {
+	if IpTables() == nil {
+		return ErrNoIptables
+	}
 	switch api {
 	case APIBanned:
+		if IpTables().Sets[IpTables().Bl] == nil {
+			return ErrNoBlacklistFound
+		}
 		return ip.Blacklist(ttl)
 	case APIAllowed:
+		if IpTables().Sets[IpTables().Wl] == nil {
+			return ErrNoWhitelistFound
+		}
 		return ip.Whitelist(ttl)
 	}
 	return fmt.Errorf("unknown API: %d", api)
@@ -127,7 +93,7 @@ func (ip *IP) Blacklist(ttl time.Duration) error {
 
 func (ip *IP) Decrypt() (string, error) {
 	if len(ip.Ipaddr) > 0 {
-		return decryptIp(ip.Ipaddr, ip.Encrypt)
+		return DecryptIp(ip.Ipaddr, ip.Encrypt)
 	}
 	return "", ErrJsonEmptyIPAddressField
 }
